# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Some desc",
        "title": "Throttler API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/throttler": {
            "post": {
                "description": "Add new request for external svc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "throttler"
                ],
                "summary": "New request",
                "operationId": "NewReq",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.newReqResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/throttler/{uuid}": {
            "get": {
                "description": "Return request by uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "throttler"
                ],
                "summary": "Request By UUID",
                "operationId": "GetRequest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.getReqResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Request": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "v1.getReqResp": {
            "type": "object",
            "properties": {
                "request": {
                    "$ref": "#/definitions/entity.Request"
                }
            }
        },
        "v1.newReqResp": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}